requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
pytesseract
pdf2image
Pillow
motor
dnspython
cloudinary
httpx
phonenumbers
google-generativeai
python-dotenv